requirements.txt
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
Flask==2.3.2
fonttools==4.39.4
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.0
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nbformat==5.8.0
nest-asyncio==1.5.6
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==3.5.0
plotly==5.14.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tenacity==8.2.2
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.3.3
yarg==0.1.9
zipp==3.15.0
